custom_tools/bbox-brush/requirements.txt
transformers
pillow